refactor: move npu cp/dp-ep padding utilities into parallel_state.
Relocate cp/ep and dp/ep padding helpers out of framework/model, then update includes and CMake deps so ownership and linkage align with parallel-state responsibilities.
