+++ xllm/core/framework/parallel_state/npu_cp_prepare.h
