+++ xllm/core/layers/npu/npu_deepseek_v2_decoder_layer_impl.h
@@ -24,8 +24,8 @@ limitations under the License.
 #include "framework/eplb/expert_buffer_manager.h"
 #include "framework/eplb/expert_weight_buffer_shm.h"
 #include "framework/model/model_input_params.h"
-#include "framework/model/npu_dp_ep_padding.h"
 #include "framework/model_context.h"
+#include "framework/parallel_state/npu_dp_ep_padding.h"
 #include "framework/state_dict/state_dict.h"
 #include "loader/deepseek_v2_decoder_loader.h"
 #include "npu_base_layer.h"
