+++ xllm/core/layers/npu/npu_glm4_moe_lite_decoder_layer.h
@@ -23,8 +23,8 @@ limitations under the License.
 
 #include "framework/model/model_args.h"
 #include "framework/model/model_input_params.h"
-#include "framework/model/npu_dp_ep_padding.h"
 #include "framework/model_context.h"
+#include "framework/parallel_state/npu_dp_ep_padding.h"
 #include "framework/quant_args.h"
 #include "framework/state_dict/state_dict.h"
 #include "loader/glm4_moe_lite_decoder_loader.h"
