+++ xllm/core/layers/npu/npu_qwen3_moe_decoder_layer_impl.h
@@ -23,7 +23,7 @@ limitations under the License.
 
 #include "framework/model/model_args.h"
 #include "framework/model/model_input_params.h"
-#include "framework/model/npu_dp_ep_padding.h"
+#include "framework/parallel_state/npu_dp_ep_padding.h"
 #include "framework/quant_args.h"
 #include "framework/state_dict/state_dict.h"
 #include "loader/qwen3_moe_decoder_loader.h"
