+++ xllm/core/runtime/worker_impl.cpp
@@ -49,8 +49,8 @@ limitations under the License.
 #include "core/distributed_runtime/master.h"
 #include "framework/kv_cache/kv_cache.h"
 #include "framework/model/model_input_params.h"
-#include "framework/model/npu_cp_ep_padding.h"
 #include "framework/model_loader.h"
+#include "framework/parallel_state/npu_cp_ep_padding.h"
 #include "framework/sampling/sampler.h"
 #include "framework/state_dict/state_dict.h"
 #include "framework/xtensor/global_xtensor.h"
