+++ xllm/models/llm/npu/glm4_moe.h
@@ -16,8 +16,8 @@ limitations under the License.
 #pragma once
 
 #include "core/framework/model/model_output.h"
-#include "core/framework/model/npu_dp_ep_padding.h"
 #include "core/framework/model_context.h"
+#include "core/framework/parallel_state/npu_dp_ep_padding.h"
 #include "core/layers/npu/npu_glm4_moe_decoder_layer.h"
 #include "llm_model_base.h"
